+++ com.unity.render-pipelines.high-definition/Runtime/Lighting/LightDefinition.cs
@@ -69,7 +69,7 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         public int     shadowIndex;             // -1 if unused (TODO: 16 bit)
 
         public Vector3 color;
-        public int     contactShadowIndex;      // -1 if unused (TODO: 16 bit)
+        public int     contactShadowMask;      // 0 if unused (TODO: 16 bit)
 
         public float   shadowDimmer;
         public float   volumetricShadowDimmer;  // Replaces 'shadowDimmer'
@@ -111,12 +111,12 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
         public int     tileCookie;              // (TODO: use a bitfield)
         public int     shadowIndex;             // -1 if unused (TODO: 16 bit)
 #if ENABLE_RAYTRACING
-        // We store the ray traced area shadow index as a negative value inside the contactShadowIndex.
+        // We store the ray traced area shadow index as a negative value inside the contactShadowMask.
         // Contact shadows are disabled for area lights and setting the index as negative allows for still
         // disabling contact shadows in the shader code (checks for => 0)
-        public int     rayTracedAreaShadowIndex { get => -contactShadowIndex; set => contactShadowIndex = -value; }
+        public int     rayTracedAreaShadowIndex { get => -contactShadowMask; set => contactShadowMask = -value; }
 #endif
-        public int contactShadowIndex;      // negative if unused (TODO: 16 bit)
+        public int     contactShadowMask;      // negative if unused (TODO: 16 bit)
 
         public float   shadowDimmer;
         public float   volumetricShadowDimmer;  // Replaces 'shadowDimmer'