+++ com.unity.visualeffectgraph/Editor/Models/Contexts/Implementations/VFXPlanarPrimitiveOutput.cs
@@ -1,0 +1,101 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using UnityEditor.VFX.Block;
+using UnityEngine;
+using UnityEngine.Experimental.VFX;
+
+namespace UnityEditor.VFX
+{
+    [VFXInfo(variantProvider = typeof(VFXPlanarPrimitiveVariantProvider))]
+    class VFXPlanarPrimitiveOutput : VFXAbstractParticleOutput
+    {
+        [VFXSetting(VFXSettingAttribute.VisibleFlags.InInspector), SerializeField]
+        protected VFXPrimitiveType primitiveType = VFXPrimitiveType.Quad;
+
+        //[VFXSetting(VFXSettingAttribute.VisibleFlags.InInspector)]
+        public bool useGeometryShader = false;
+
+        public override string name { get { return primitiveType.ToString() + " Output"; } }
+        public override string codeGeneratorTemplate { get { return RenderPipeTemplate("VFXParticlePlanarPrimitive"); } }
+        public override VFXTaskType taskType
+        {
+            get
+            {
+                if (useGeometryShader)
+                    return VFXTaskType.ParticlePointOutput;
+
+                return VFXPlanarPrimitiveHelper.GetTaskType(primitiveType);
+            }
+        }
+        public override bool supportsUV { get { return true; } }
+
+        public override IEnumerable<string> additionalDefines
+        {
+            get
+            {
+                foreach (var def in base.additionalDefines)
+                    yield return def;
+
+                if (useGeometryShader)
+                    yield return "USE_GEOMETRY_SHADER";
+
+                yield return VFXPlanarPrimitiveHelper.GetShaderDefine(primitiveType);
+            }
+        }
+
+        public override IEnumerable<VFXAttributeInfo> attributes
+        {
+            get
+            {
+                yield return new VFXAttributeInfo(VFXAttribute.Position, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.Color, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.Alpha, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.Alive, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.AxisX, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.AxisY, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.AxisZ, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.AngleX, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.AngleY, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.AngleZ, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.PivotX, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.PivotY, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.PivotZ, VFXAttributeMode.Read);
+
+                yield return new VFXAttributeInfo(VFXAttribute.Size, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.ScaleX, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.ScaleY, VFXAttributeMode.Read);
+                yield return new VFXAttributeInfo(VFXAttribute.ScaleZ, VFXAttributeMode.Read);
+
+                if (usesFlipbook)
+                    yield return new VFXAttributeInfo(VFXAttribute.TexIndex, VFXAttributeMode.Read);
+            }
+        }
+
+        protected override IEnumerable<VFXPropertyWithValue> inputProperties
+        {
+            get
+            {
+                var properties = base.inputProperties;
+                if (primitiveType == VFXPrimitiveType.Octagon)
+                    properties = properties.Concat(PropertiesFromType(typeof(VFXPlanarPrimitiveHelper.OctagonInputProperties)));
+                return properties;
+            }
+        }
+
+        protected override IEnumerable<VFXNamedExpression> CollectGPUExpressions(IEnumerable<VFXNamedExpression> slotExpressions)
+        {
+            foreach (var exp in base.CollectGPUExpressions(slotExpressions))
+                yield return exp;
+
+            yield return slotExpressions.First(o => o.name == "mainTexture");
+            if (primitiveType == VFXPrimitiveType.Octagon)
+                yield return slotExpressions.First(o => o.name == "cropFactor");
+        }
+
+        public class InputProperties
+        {
+            public Texture2D mainTexture = VFXResources.defaultResources.particleTexture;
+        }
+    }
+}